Moving the project to a module-based approach. Adding requirements.txt, pytest and setup.py
