--- main.py
@@ -1,32 +1,0 @@
-#!/usr/bin/env python2
-"""
-BORICE is software for the estimation of population mean outcrossing rates and
-inbreeding coefficients using Bayesian methods.
-Copyright (C) 2012 Vanessa A. Koelling, Patrick J. Monnahan, John K. Kelly
-
-This program is free software: you can redistribute it and/or modify it under 
-the terms of the GNU General Public License as published by the Free Software Foundation,
-either version 3 of the License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>
-"""
-from PyQt5 import QtCore, QtWidgets
-import sys
-import BORICE_GUI
-
-def main():
-	app = QtWidgets.QApplication(sys.argv)
-	window = BORICE_GUI.MainWindow(None, app)
-	window.show()
-	sys.exit(app.exec_())
-
-if __name__ == '__main__': 
-	main()
-
-
