+++ requirements.txt
@@ -1,0 +1,3 @@
+PyQt5==5.15.7
+pytest==7.1.2
+setuptools==58.1.0
