+++ tests/test_borice.py
@@ -1,0 +1,5 @@
+from borice.application import *
+
+# Test borice with default parameters
+def test_borice():
+    Application().run("example_datafile.csv", [True, True, True])
